requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Templates
jinja2==3.1.3

# HTTP client
httpx==0.26.0
requests==2.31.0

# Production server
gunicorn==21.2.0

# Financial calculations
numpy==1.26.3
pandas==2.1.4

# Date handling
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Forms
python-multipart==0.0.6
